// frontend/src/app/models/Student.model.ts
export interface Student {
  _id?: string;
  fisrtName?: string;
  lastName?: string;
  StudentNumber?: string;
  email?: string;
  password?: string;
  address?: string;
  city?: string;
  program?: string;
  phone?: string;
  salt?: string;
  provider?: string;
  providerId?:string;
  providerData?: string;
  created?: Date;

}